mutli-agent-system/ai_homework-main/ai_homework-main/muti-agent-system/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.55.3
python-dotenv==1.0.0
requests==2.31.0
pydantic>=2.0.0
nltk==3.8.1
